Add elo and better seeding logic

+++ critic/src/dto/category_item.rs
@@ -2,7 +2,7 @@ pub struct CategoryItem {
     pub id: i32,
     pub name: String,
     pub sub_categories: Vec<String>,
-    pub elo: i32,
+    pub elo: f32,
 }
 
 pub struct NewCategoryItem {
