+++ critic/src/dto/contest.rs
@@ -3,7 +3,7 @@ use super::Category;
 pub struct Contestant {
     pub id: i32,
     pub name: String,
-    pub elo: i32,
+    pub elo: f32,
 }
 
 pub struct Contest {
